Bugfix: Wrong usage of dim_vector.numel() to determine rank of array.
From the documentation of dim_vector.numel() in GNU Octave 5.2.0:
Return the number of elements that a matrix with this dimension vector
would have, NOT the number of dimensions (elements in the dimension
vector).

The correct call is dim_vector.ndims().

Consequently, the template alloc_hsize(...) was removed and replaced by
separate versions for Matrix and dim_vector data types.

Now all tests pass.